fix: replace to Head API #525
