+++ gatsby-ssr.tsx
@@ -1,15 +1,27 @@
 import { getInitColorSchemeScript } from "@mui/material/styles";
-import { GatsbySSR } from "gatsby";
-import React from "react";
+import { Fragment } from "react";
+
+import type { GatsbySSR } from "gatsby";
+
+import { WrapPageElement } from "@/layouts/WrapPageElement";
+import { WrapRootElement } from "@/layouts/WrapRootElement";
 
 export const onRenderBody: GatsbySSR["onRenderBody"] = ({
   setPreBodyComponents,
 }) => {
   setPreBodyComponents([
-    <React.Fragment key="init-color-scheme-script">
+    <Fragment key="init-color-scheme-script">
       {getInitColorSchemeScript({
         enableSystem: true,
       })}
-    </React.Fragment>,
+    </Fragment>,
   ]);
+};
+
+export const wrapRootElement: GatsbySSR['wrapRootElement'] = ({ element }) => {
+  return <WrapRootElement>{element}</WrapRootElement>;
+};
+
+export const wrapPageElement: GatsbySSR['wrapPageElement'] = ({ element }) => {
+  return <WrapPageElement>{element}</WrapPageElement>;
 };
