+++ src/components/index.ts
@@ -17,10 +17,3 @@ export * from "./organisms/ProjectList";
 export * from "./organisms/RelatedBlogPostList";
 export * from "./organisms/SkillList";
 export * from "./organisms/WhatICanDoList";
-export * from "./templates/BlogPost";
-export * from "./templates/CookieAlert";
-export * from "./templates/Footer";
-export * from "./templates/Head";
-export * from "./templates/Header";
-export * from "./templates/Layout";
-export * from "./templates/TopLayout";
