+++ src/locales/ja/translation.json
@@ -19,6 +19,7 @@
     "changeLanguage-title": "言語切替",
     "github-title": "GitHub",
     "github-url": "https://github.com/bicstone/",
+    "icon": "ニコニコした Oishi Takanori の似顔絵",
     "title-home": "Oishi Takanori",
     "toggleDarkTheme-title": "ダークモード切替"
   },
