+++ src/pages/404.tsx
@@ -1,48 +1,34 @@
-import { SelfImprovement as SelfImprovementIcon } from "@mui/icons-material";
-import { Typography, Container, Button } from "@mui/material";
-import { graphql, PageProps, Link as RouterLink } from "gatsby";
-import { useI18next } from "gatsby-plugin-react-i18next";
-import React from "react";
-import { Layout } from "src/components";
-import { NotFoundPageQuery } from "src/types";
+import { SelfImprovement as SelfImprovementIcon } from '@mui/icons-material';
+import { Typography, Container, Button } from '@mui/material';
+import { graphql, PageProps, Link as RouterLink, HeadFC } from 'gatsby';
+import { useI18next } from 'gatsby-plugin-react-i18next';
+import React from 'react';
+import { NotFoundPageQuery } from 'src/types';
 
-const NotFound: React.FC<PageProps<NotFoundPageQuery>> = ({ data }) => {
+import { HeadTemplate } from '@/templates/HeadTemplate';
+
+const NotFound: React.FC<PageProps<NotFoundPageQuery>> = () => {
   const { t } = useI18next();
-  const icon = data.icon.svg.content;
-  const iconAlt = data.icon.title;
+
   return (
-    <Layout icon={icon} iconAlt={iconAlt}>
-      <Container maxWidth="md">
-        <div
-          css={(theme) => ({ margin: theme.spacing(2), textAlign: "center" })}
-        >
-          {/* 見つかりませんでした */}
-          <SelfImprovementIcon
-            css={(theme) => ({
-              width: theme.spacing(20),
-              height: theme.spacing(20),
-            })}
-          />
-          <Typography variant="h4" component="h1" paragraph>
-            {t("not-found.title")}
-          </Typography>
-          <Typography variant="body1">{t("not-found.description")}</Typography>
-        </div>
-        <div
-          css={(theme) => ({ margin: theme.spacing(2), textAlign: "center" })}
-        >
-          {/* ホームに戻る */}
-          <Button
-            component={RouterLink}
-            variant="contained"
-            to="/"
-            size="large"
-          >
-            {t("not-found.back-to-home")}
-          </Button>
-        </div>
-      </Container>
-    </Layout>
+    <Container maxWidth="md">
+      <div css={theme => ({ margin: theme.spacing(2), textAlign: 'center' })}>
+        {/* 見つかりませんでした */}
+        <SelfImprovementIcon
+          css={theme => ({ width: theme.spacing(20), height: theme.spacing(20) })}
+        />
+        <Typography variant="h4" component="h1" paragraph>
+          {t('not-found.title')}
+        </Typography>
+        <Typography variant="body1">{t('not-found.description')}</Typography>
+      </div>
+      <div css={theme => ({ margin: theme.spacing(2), textAlign: 'center' })}>
+        {/* ホームに戻る */}
+        <Button component={RouterLink} variant="contained" to="/" size="large">
+          {t('not-found.back-to-home')}
+        </Button>
+      </div>
+    </Container>
   );
 };
 
@@ -70,3 +56,8 @@ export const query = graphql`
     }
   }
 `;
+
+export const Head: HeadFC = () => {
+  return <HeadTemplate />;
+};
+
