+++ tsconfig.json
@@ -49,6 +49,8 @@
     "baseUrl": "./" /* Base directory to resolve non-absolute module names. */,
     "paths": {
       "@/commons/*": ["src/commons/*"],
+      "@/layouts/*": ["src/layouts/*"],
+      "@/templates/*": ["src/templates/*"],
       "src": ["src/*"]
     } /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */,
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
